+++ CHANGELOG.md
@@ -3,10 +3,16 @@
 ## [1.5.0](https://github.com/XzantGaming/Repak-X/releases/latest)
 
 ### 🔧 Backend / Logic
+- UAssetTool communication rewrite for a more stable and faster connection
 - Added detection for Hybrid IO Store Mods (Mods containing both raw and uasset files)
 
 ### 🎨 Frontend / UI
+- Added a Sig Bypasser toggle in the Tools panel
+- Mod type labels are now additive
+- Dropping mods (pak/folders) into Mod Install panel will now add them to the installation queue
+- Pressing 'Enter' in the Mod Install panel will now execute the installation queue
 - Added Hybrid label for Hybrid IO Store Mods
 - Added sorting for mods by modified date and name with asc/desc order
 - Added a resize handle for the filters and folders panel
 - Added Cyclops hero icon
+- Minor UI changes